Budget amounts are not getting saved in the glbudgetdetails table This fix inserts a zero balance for al;l periods/accounts when a new budget is created
 Files changed in commit:
 GLBudgetHeaders.php
 sql/updates/7.php

+++ sql/updates/7.php
@@ -9,8 +9,8 @@ CreateTable('glbudgetheaders', "CREATE TABLE IF NOT EXISTS `glbudgetheaders` (
   `owner` varchar(20) NOT NULL DEFAULT '',
   `name` varchar(200) NOT NULL DEFAULT '',
   `description` text,
-  `startperiod` smallint(6) NOT NULL DEFAULT 0,
-  `endperiod` smallint(6) NOT NULL DEFAULT 0,
+  `startperiod` INT NOT NULL DEFAULT 0,
+  `endperiod` INT NOT NULL DEFAULT 0,
   `current` tinyint(1) NOT NULL DEFAULT 0,
   PRIMARY KEY  (`id`)
 )");
@@ -19,7 +19,7 @@ CreateTable('glbudgetdetails', "CREATE TABLE IF NOT EXISTS `glbudgetdetails` (
   `id` int(11) NOT NULL auto_increment,
   `headerid` int(11) NOT NULL DEFAULT 0,
   `account` varchar(20) NOT NULL DEFAULT '',
-  `period` smallint(6) NOT NULL DEFAULT 0,
+  `period` INT NOT NULL DEFAULT 0,
   `amount` double NOT NULL DEFAULT 0.0,
   PRIMARY KEY  (`id`),
   KEY (`account`),
